requirement.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
sentence-transformers
faiss-cpu
langchain-community
pypdf
